stock_skills/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
yfinance==0.2.38
pyyaml==6.0.1
pydantic==2.7.1
